+++ CHANGELOG.rst
@@ -3,6 +3,21 @@ chronological order.
 
 
 
+0.15.1 (Monday 8th September 2025)
+----------------------------------
+
+
+Changed
+^^^^^^^
+
+* Switched to ``setuptools-scm`` for version number management.
+
+
+Fixed
+^^^^^
+
+* Fixed a refresh issue in the :class:`.EditableListBox`.
+
 
 0.15.0 (Wednesday 5th February 2025)
 ------------------------------------